GH-576 removed unneeded subclassing
