+++ SafeUIKit/SafeAppUI/AppFlow/MainFlow/Menu/ConnectBrowserExtension/ConnectBrowserExtensionLaterCommand.swift
@@ -7,8 +7,6 @@ import MultisigWalletApplication
 
 final class ConnectBrowserExtensionLaterCommand: MenuCommand {
 
-    private var commandFlow = ConnectBrowserExtensionFlowCoordinator()
-
     override var title: String {
         return LocalizedString("menu.action.connect_browser_extension",
                                comment: "Connect browser extension menu item.")
@@ -18,17 +16,9 @@ final class ConnectBrowserExtensionLaterCommand: MenuCommand {
         return !ApplicationServiceRegistry.connectExtensionService.isAvailable
     }
 
-    override func run(mainFlowCoordinator: MainFlowCoordinator) {
-        mainFlowCoordinator.saveCheckpoint()
-        mainFlowCoordinator.enter(flow: commandFlow) { [unowned mainFlowCoordinator] in
-            DispatchQueue.main.async {
-                mainFlowCoordinator.popToLastCheckpoint()
-                mainFlowCoordinator.pop()
-                DispatchQueue.main.asyncAfter(deadline: .now() + .milliseconds(800)) {
-                    mainFlowCoordinator.showTransactionList()
-                }
-            }
-        }
+    override init() {
+        super.init()
+        childFlowCoordinator = ConnectBrowserExtensionFlowCoordinator()
     }
 
 }
