+++ SafeUIKit/SafeAppUI/AppFlow/MainFlow/Menu/ReplaceBrowserExtension/ReplaceBrowserExtensionCommand.swift
@@ -16,19 +16,9 @@ class ReplaceBrowserExtensionCommand: MenuCommand {
         return !ApplicationServiceRegistry.replaceExtensionService.isAvailable
     }
 
-    var commandFlow = ReplaceBrowserExtensionFlowCoordinator()
-
-    override func run(mainFlowCoordinator: MainFlowCoordinator) {
-        mainFlowCoordinator.saveCheckpoint()
-        mainFlowCoordinator.enter(flow: commandFlow) { [unowned mainFlowCoordinator] in
-            DispatchQueue.main.async {
-                mainFlowCoordinator.popToLastCheckpoint()
-                mainFlowCoordinator.pop()
-                DispatchQueue.main.asyncAfter(deadline: .now() + .milliseconds(800)) {
-                    mainFlowCoordinator.showTransactionList()
-                }
-            }
-        }
+    override init() {
+        super.init()
+        childFlowCoordinator = ReplaceBrowserExtensionFlowCoordinator()
     }
 
 }
